various script changes

+++ mc-mapping.c
@@ -218,7 +218,7 @@ int main(int argc, char* argv[])
 
 	printf("size: %d (%d KB)\n", g_mem_size, g_mem_size/1024);
 	printf("duration %.0f ns, #access %d\n", (double)nsdiff, naccess);
-	printf("average latency: %ld ns\n", nsdiff/naccess);
+	printf("average latency: %.0f ns\n", (double)nsdiff/naccess);
 	printf("bandwidth %.2f MB/s\n", (double)64*1000*naccess/nsdiff);
 
 	return 0;
